backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
pillow==10.4.0
numpy==1.26.4
torch
torchvision
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
PyMuPDF==1.24.11
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
python-dotenv==1.0.1
boto3==1.35.24
gspread==6.1.2
google-auth==2.35.0
